+++ test/test_LensModel/test_single_plane.py
@@ -45,8 +45,8 @@ class TestLensModel(object):
         npt.assert_almost_equal(density, density_model, decimal=8)
 
     def test_bool_list(self):
-        lensModel = SinglePlane(['SPEMD', 'SHEAR'])
-        kwargs = [{'theta_E': 1, 'gamma': 1, 'e1': 0.1, 'e2': -0.1, 'center_x': 0, 'center_y': 0},
+        lensModel = SinglePlane(['SPEP', 'SHEAR'])
+        kwargs = [{'theta_E': 1, 'gamma': 2, 'e1': 0.1, 'e2': -0.1, 'center_x': 0, 'center_y': 0},
                            {'gamma1': 0.01, 'gamma2': -0.02}]
         alphax_1, alphay_1 = lensModel.alpha(1, 1, kwargs, k=0)
         alphax_1_list, alphay_1_list = lensModel.alpha(1, 1, kwargs, k=[0])
@@ -60,7 +60,7 @@ class TestLensModel(object):
         npt.assert_almost_equal(alphay_1_1 + alphay_1_2, alphay_full, decimal=5)
 
     def test_init(self):
-        lens_model_list = ['TNFW', 'SPEMD_SMOOTH', 'TRIPLE_CHAMELEON', 'SHEAR_GAMMA_PSI', 'CURVED_ARC', 'NFW_MC',
+        lens_model_list = ['TNFW', 'PEMD', 'SPEMD', 'TRIPLE_CHAMELEON', 'SHEAR_GAMMA_PSI', 'CURVED_ARC', 'NFW_MC',
                            'ARC_PERT']
         lensModel = SinglePlane(lens_model_list=lens_model_list)
         assert lensModel.func_list[0].param_names[0] == 'Rs'
